Move styling settings to inner blocks
