+++ plugins/woocommerce-blocks/assets/js/blocks/product-filters/inner-blocks/active-filters/edit.tsx
@@ -5,12 +5,7 @@ import {
 	useBlockProps,
 	useInnerBlocksProps,
 	BlockContextProvider,
-	BlockControls,
 } from '@wordpress/block-editor';
-import { ToolbarGroup, ToolbarButton } from '@wordpress/components';
-import { arrowRight, arrowDown } from '@wordpress/icons';
-import { __ } from '@wordpress/i18n';
-import { getBlockSupport } from '@wordpress/blocks';
 
 /**
  * Internal dependencies
@@ -23,13 +18,8 @@ import { EditProps } from './types';
 import { filtersPreview } from './constants';
 
 const Edit = ( props: EditProps ) => {
-	const { attributes, setAttributes, name } = props;
-	const { clearButton, layout } = attributes;
-
-	// Extract attributes from block layout
-	const layoutBlockSupport = getBlockSupport( name, 'layout' );
-	const defaultBlockLayout = layoutBlockSupport?.default;
-	const usedLayout = layout || defaultBlockLayout || {};
+	const { attributes } = props;
+	const { clearButton } = attributes;
 
 	const { children, ...innerBlocksProps } = useInnerBlocksProps(
 		useBlockProps(),
@@ -65,39 +55,6 @@ const Edit = ( props: EditProps ) => {
 	return (
 		<div { ...innerBlocksProps }>
 			<Inspector { ...props } />
-			<BlockControls>
-				<ToolbarGroup>
-					<ToolbarButton
-						icon={ arrowRight }
-						label={ __( 'Horizontal', 'woocommerce' ) }
-						onClick={ () =>
-							setAttributes( {
-								layout: {
-									...usedLayout,
-									orientation: 'horizontal',
-								},
-							} )
-						}
-						isPressed={
-							usedLayout.orientation === 'horizontal' ||
-							! usedLayout.orientation
-						}
-					/>
-					<ToolbarButton
-						icon={ arrowDown }
-						label={ __( 'Vertical', 'woocommerce' ) }
-						onClick={ () =>
-							setAttributes( {
-								layout: {
-									...usedLayout,
-									orientation: 'vertical',
-								},
-							} )
-						}
-						isPressed={ usedLayout.orientation === 'vertical' }
-					/>
-				</ToolbarGroup>
-			</BlockControls>
 			<InitialDisabled>
 				<BlockContextProvider
 					value={ {
