+++ plugins/woocommerce-blocks/assets/js/blocks/product-filters/inner-blocks/active-filters/types.ts
@@ -5,11 +5,6 @@ import { BlockEditProps } from '@wordpress/blocks';
 
 export type BlockAttributes = {
 	clearButton: boolean;
-	layout: {
-		orientation: string;
-	};
 };
 
-export type EditProps = BlockEditProps< BlockAttributes > & {
-	name: string;
-};
+export type EditProps = BlockEditProps< BlockAttributes >;
