+++ plugins/woocommerce-blocks/assets/js/blocks/product-filters/inner-blocks/removable-chips/block.json
@@ -13,7 +13,17 @@
 	"ancestor": [
 		"woocommerce/product-filter-active"
 	],
-	"supports": {},
+	"supports": {
+		"layout": {
+			"allowSwitching": false,
+			"allowInheriting": false,
+			"allowJustification": false,
+			"allowVerticalAlignment": false,
+			"default": {
+				"type": "flex"
+			}
+		}
+	},
 	"usesContext": [
 		"queryId",
 		"filterData"
